Support JWT 2.0 and multiple algorithms
Since JWT 2.0 the signature verification checks that the algorithm
passed to `encode` is the same as the `alg` value in the token. If you
support multiple algorithms you can pass an `algorithms: []` option to
the `JWT.decode` call.

This updates the omniauth strategy to allow us passing multiple options
to the `decode` call to support multiple algorithms (and not just
HS256).

* https://github.com/jwt/ruby-jwt/pull/184